'Middleware-y-cambio-carpetas-controladores'

+++ routes/web.php
@@ -1,6 +1,9 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\Admin\DoctorController;
+use App\Http\Controllers\Admin\PatientController;
+use App\Http\Middleware\MiMiddlewarePersonalizado;
 
 Route::get('/', function () {
     return view('welcome');
@@ -10,6 +13,11 @@ Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
+Route::middleware(['auth','admin'])->group(function () {
+
+//namespace busca en la carpeta que se le indique    
+//Route::middleware(['auth','admin'])->namespace('Admin')->group(function () {
+
 //Antiguo
 //Route::get('/home', 'HomeController@index')->name('home');
 
@@ -17,30 +25,32 @@ Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name
 
 //Antiguo
 //Route::get('/specialties', 'SpecialtyController@index');
-Route::get('/specialties', [\App\Http\Controllers\SpecialtyController::class, 'index']);
+Route::get('/specialties', [\App\Http\Controllers\Admin\SpecialtyController::class, 'index']);
 
 //Route::get('/specialties/create', 'SpecialtyController@create');//form registro
-Route::get('/specialties/create', [\App\Http\Controllers\SpecialtyController::class, 'create']);
+Route::get('/specialties/create', [\App\Http\Controllers\Admin\SpecialtyController::class, 'create']);
 
 
 //Route::get('/specialties/{specialty}/edit', 'SpecialtyController@edit');
-Route::get('/specialties/{specialty}/edit', [\App\Http\Controllers\SpecialtyController::class, 'edit']);
+Route::get('/specialties/{specialty}/edit', [\App\Http\Controllers\Admin\SpecialtyController::class, 'edit']);
 
 //Route::get('/specialties', 'SpecialtyController@store');// envio del form
-Route::post('/specialties', [\App\Http\Controllers\SpecialtyController::class, 'store']);
+Route::post('/specialties', [\App\Http\Controllers\Admin\SpecialtyController::class, 'store']);
 
-Route::put('/specialties/{specialty}', [\App\Http\Controllers\SpecialtyController::class, 'update']);
+Route::put('/specialties/{specialty}', [\App\Http\Controllers\Admin\SpecialtyController::class, 'update']);
 
-Route::delete('/specialties/{specialty}', [\App\Http\Controllers\SpecialtyController::class, 'destroy']);
+Route::delete('/specialties/{specialty}', [\App\Http\Controllers\Admin\SpecialtyController::class, 'destroy']);
 
 
 //Doctors
 
 //Route::resource('doctors','DoctorController');
-use App\Http\Controllers\DoctorController;
+
 Route::resource('/doctors', DoctorController::class);
 
 
 //Patients
-use App\Http\Controllers\PatientController;
+
 Route::resource('/patients', PatientController::class);
+
+});
